(QENG-1192) Use hostname specified by hypervisor everywhere.
Previously, we would call out to `hostname` on the SUT when deploying
PE. This may give us a name that isn't resolvable our routable by
every host in the configuration. We should instead just use whatever
the hypervisor tells us is the correct way to connect to the host.

This also sneaks in a small change to the vmpooler hypervisor, to make
it return FQDNs. Using these FQDNs is really what this is all about.