(QENG-1264) Allow users to specify recursive option for scp_{from,to}
The fix for QENG-1128 caused scp_from and scp_to to ignore the
recursive flag. This caused all scp_from operations to be recursive,
which will cause a failure if trying to copy only a single file.

The fix to this, without regressing on QENG-1128, is to actually copy
the options hash locally so that we can modify it as needed for
defaults without updating any global data.